agent_python/requirements.txt
Flask
Flask-SocketIO
psutil
gevent
py-cpuinfo
gputil
pyyaml
Flask-WTF
python-dotenv
Flask-Cors